updated datagen
